validations

+++ app/Http/Controllers/CategoryController.php
@@ -8,24 +8,14 @@ use Illuminate\Support\Facades\DB;
 
 class CategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
-        $categories = Category::paginate(5);
+        $categories = Category::paginate(10);
         return view('categories.index', [
             'categories' => $categories,
         ]);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         $allCategories = Category::all();
@@ -39,42 +29,27 @@ class CategoryController extends Controller
         ]);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
+        $request->validate([
+            'name_en' => 'required|min:3',
+        ]);
         $request['status'] = isset($request->status) ? 1 : 0;
         Category::create($request->all());
         return redirect('/categories');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
     public function show(Category $category)
     {
         return view('categories.view', ['category' => $category]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Category $category)
     {
         $allCategories = Category::where('id', '!=', $category->id)->get();
         $parentsCategories = [];
         foreach ($allCategories as $oneCategory) {
-            if (!isset($oneCategory->parent->parent))
+            if (!isset($oneCategory->parent->parent) && !in_array($oneCategory->id, $category->children->modelKeys()))
                 $parentsCategories[] = $oneCategory;
         }
         return view('categories.edit', [
@@ -83,29 +58,20 @@ class CategoryController extends Controller
         ]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, Category $category)
     {
         $request['status'] = isset($request->status) ? 1 : 0;
+        $request->validate([
+            'name_en' => 'required|min:3',
+        ]);
         $category->update($request->all());
         return redirect('/categories');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(Category $category)
     {
-        $category->delete();
+        if (!json_decode($category->children) && !json_decode($category->offers))
+            $category->delete();
         return redirect('/categories');
     }
 
@@ -113,14 +79,6 @@ class CategoryController extends Controller
     {
         $category->status = $category->status ? 0 : 1;
         $category->update();
-        // foreach ($category->children as $category) {
-        //     $category->status = $category->status ? 0 : 1;
-        //     $category->update();
-        //     foreach ($category->children as $category) {
-        //         $category->status = $category->status ? 0 : 1;
-        //         $category->update();
-        //     }
-        // }
         return redirect('/categories');
     }
